websocket-main_archived/requirements.txt
websockets
paho-mqtt
git+https://github.com/project56HR/DatabaseManager.git
ruff